ref(prettier): extract rule warning against competing configurations
