+++ src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './package-json';
 export * from './rules/install-dependencies';
 export * from './rules/install-options';
+export * from './rules/warn-against-competing-configuration';
