+++ src/lib/rules/competing-configuration-options.ts
@@ -1,0 +1,4 @@
+export interface CompetingConfigurationOptions {
+  packageJsonKey: string;
+  files: string[];
+}
